feat: improve TypeScript type safety and extract BigFivePentagon component
- Replace `any` types with proper TypeScript interfaces for assessment results
- Extract BigFivePentagon component from inline definition to separate file
- Add comprehensive type definitions for BigFive, MBTI, and DISC results
- Improve type safety throughout renderInlineResults function
- Add accessibility attributes to pentagon chart (ARIA labels, screen reader support)

Addresses high priority concerns from code review to improve maintainability and type safety.
