+++ src/screens/PersonalityAssessments/PersonalityAssessmentOverlay.tsx
@@ -29,22 +29,11 @@ import { cn } from "@/lib/utils"
 import BigFiveResults from "./components/BigFiveResults"
 import MBTIResults from "./components/MBTIResults"
 import DISCResults from "./components/DISCResults"
+import { BigFivePentagon } from "./components/BigFivePentagon"
 import { routes } from "@/firebaseConfig"
 import { db } from "@/lib/firebase"
 import { doc, setDoc, serverTimestamp, onSnapshot } from "firebase/firestore"
-
-interface PersonalityAssessment {
-  assessment_id: string
-  session_id: string
-  name: string
-  description: string
-  completed_at: string | null
-  results: any
-  answers: any
-  questions_answered: number
-  started_at: number
-  completed: boolean
-}
+import { PersonalityAssessment, BigFiveResults as BigFiveResultsType, MBTIResults as MBTIResultsType, DISCResults as DISCResultsType } from "./types/assessmentTypes"
 
 interface LastSyncStatus {
   can_sync: boolean
@@ -454,126 +443,21 @@ export default function PersonalityAssessmentOverlay() {
     }
   }
 
-  // Pentagon radar chart component for Big Five
-  const BigFivePentagon = ({ results }: { results: any }) => {
-    const dimensions = Object.entries(results.dimension_scores || {})
-    const size = 120
-    const center = size / 2
-    const radius = 35
-
-    const getPolygonPoints = (scores: number[]) => {
-      return scores
-        .map((score, index) => {
-          const angle = (index * 2 * Math.PI) / 5 - Math.PI / 2
-          const value = (score / 5) * radius
-          const x = center + value * Math.cos(angle)
-          const y = center + value * Math.sin(angle)
-          return `${x},${y}`
-        })
-        .join(" ")
-    }
-
-    const getOutlinePoints = () => {
-      const points = []
-      for (let i = 0; i < 5; i++) {
-        const angle = (i * 2 * Math.PI) / 5 - Math.PI / 2
-        const x = center + radius * Math.cos(angle)
-        const y = center + radius * Math.sin(angle)
-        points.push(`${x},${y}`)
-      }
-      return points.join(" ")
-    }
-
-    const getDimensionLabel = (key: string) => {
-      const labels: { [key: string]: string } = {
-        openness: "O",
-        conscientiousness: "C",
-        extraversion: "E",
-        agreeableness: "A",
-        neuroticism: "N",
-      }
-      return labels[key] || key.charAt(0).toUpperCase()
-    }
-
-    const scores = dimensions.map(([_, dim]: [string, any]) => dim.score)
-
-    return (
-      <div className="flex flex-col items-center gap-2">
-        <svg width={size} height={size} className="drop-shadow-sm">
-          {/* Grid lines */}
-          <polygon
-            points={getOutlinePoints()}
-            fill="none"
-            stroke="rgb(203 213 225)"
-            strokeWidth="1"
-            className="opacity-30"
-          />
-
-          {/* Data polygon */}
-          <polygon
-            points={getPolygonPoints(scores)}
-            fill="rgb(59 130 246)"
-            fillOpacity="0.3"
-            stroke="rgb(59 130 246)"
-            strokeWidth="2"
-          />
-
-          {/* Dimension labels */}
-          {dimensions.map(([key, dim]: [string, any], index) => {
-            const angle = (index * 2 * Math.PI) / 5 - Math.PI / 2
-            const labelRadius = radius + 15
-            const x = center + labelRadius * Math.cos(angle)
-            const y = center + labelRadius * Math.sin(angle)
-
-            return (
-              <text
-                key={key}
-                x={x}
-                y={y}
-                textAnchor="middle"
-                dominantBaseline="middle"
-                className="text-xs font-medium fill-current"
-              >
-                {getDimensionLabel(key)}
-              </text>
-            )
-          })}
-
-          {/* Center dot */}
-          <circle cx={center} cy={center} r="2" fill="rgb(59 130 246)" />
-        </svg>
-
-        <div className="flex flex-wrap gap-1 justify-center text-xs">
-          {dimensions.map(([key, dim]: [string, any]) => (
-            <Badge
-              key={key}
-              variant="outline"
-              className="text-xs px-1.5 py-0.5"
-            >
-              {getDimensionLabel(key)}: {dim.score.toFixed(1)}
-            </Badge>
-          ))}
-        </div>
-      </div>
-    )
-  }
 
   const renderInlineResults = (assessment: PersonalityAssessment) => {
     switch (assessment.assessment_id) {
-      case "big-five":
+      case "big-five": {
+        const bigFiveResults = assessment.results as BigFiveResultsType
         return (
           <div className="flex items-center justify-between">
-            <BigFivePentagon results={assessment.results} />
+            <BigFivePentagon results={bigFiveResults} />
             <div className="flex-1 ml-4">
               <h4 className="font-medium text-sm mb-2">Top Traits</h4>
               <div className="space-y-1">
-                {Object.entries(assessment.results?.dimension_scores || {})
-                  .sort(
-                    ([, a]: [string, any], [, b]: [string, any]) =>
-                      b.score - a.score
-                  )
+                {Object.entries(bigFiveResults.dimension_scores || {})
+                  .sort(([, a], [, b]) => b.score - a.score)
                   .slice(0, 3)
-                  .map(([key, dim]: [string, any]) => (
+                  .map(([key, dim]) => (
                     <div
                       key={key}
                       className="flex items-center justify-between text-sm"
@@ -588,8 +472,10 @@ export default function PersonalityAssessmentOverlay() {
             </div>
           </div>
         )
+      }
 
-      case "mbti":
+      case "mbti": {
+        const mbtiResults = assessment.results as MBTIResultsType
         return (
           <div className="text-center">
             <div className="mb-3">
@@ -597,12 +483,12 @@ export default function PersonalityAssessmentOverlay() {
                 variant="outline"
                 className="text-2xl font-bold py-2 px-4 bg-purple-100 text-purple-800 border-purple-300 dark:bg-purple-900/20 dark:text-purple-300 dark:border-purple-700"
               >
-                {assessment.results?.personality_type || "N/A"}
+                {mbtiResults.personality_type || "N/A"}
               </Badge>
             </div>
             <div className="grid grid-cols-2 gap-2 text-xs">
-              {Object.entries(assessment.results?.dimension_details || {}).map(
-                ([key, details]: [string, any]) => (
+              {Object.entries(mbtiResults.dimension_details || {}).map(
+                ([key, details]) => (
                   <div key={key} className="flex items-center justify-between">
                     <span className="uppercase">{key}</span>
                     <Badge variant="outline" className="text-xs">
@@ -614,8 +500,10 @@ export default function PersonalityAssessmentOverlay() {
             </div>
           </div>
         )
+      }
 
-      case "disc":
+      case "disc": {
+        const discResults = assessment.results as DISCResultsType
         return (
           <div className="space-y-3">
             <div className="flex items-center justify-center gap-2">
@@ -624,19 +512,19 @@ export default function PersonalityAssessmentOverlay() {
                 className="bg-green-100 text-green-800 border-green-300 dark:bg-green-900/20 dark:text-green-300 dark:border-green-700"
               >
                 Primary:{" "}
-                {assessment.results?.primary_style?.id.toUpperCase() || "N/A"}
+                {discResults.primary_style?.id.toUpperCase() || "N/A"}
               </Badge>
               <Badge
                 variant="outline"
                 className="bg-blue-100 text-blue-800 border-blue-300 dark:bg-blue-900/20 dark:text-blue-300 dark:border-blue-700"
               >
                 Secondary:{" "}
-                {assessment.results?.secondary_style?.id.toUpperCase() || "N/A"}
+                {discResults.secondary_style?.id.toUpperCase() || "N/A"}
               </Badge>
             </div>
             <div className="grid grid-cols-2 gap-2 text-xs">
-              {Object.entries(assessment.results?.dimension_scores || {}).map(
-                ([key, dim]: [string, any]) => (
+              {Object.entries(discResults.dimension_scores || {}).map(
+                ([key, dim]) => (
                   <div key={key} className="flex items-center justify-between">
                     <span className="uppercase">{key}</span>
                     <Badge variant="outline" className="text-xs">
@@ -648,6 +536,7 @@ export default function PersonalityAssessmentOverlay() {
             </div>
           </div>
         )
+      }
 
       default:
         return (
